+++ src/main/java/com/board_of_ads/service/impl/UserServiceImpl.java
@@ -4,11 +4,15 @@ import com.board_of_ads.models.Image;
 import com.board_of_ads.models.User;
 import com.board_of_ads.repository.ImageRepository;
 import com.board_of_ads.repository.UserRepository;
+import com.board_of_ads.service.interfaces.AuthorizationService;
 import com.board_of_ads.service.interfaces.RoleService;
 import com.board_of_ads.service.interfaces.UserService;
+import com.board_of_ads.util.BindingResultLogs;
 import lombok.AllArgsConstructor;
+import org.slf4j.Logger;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.validation.BindingResult;
 
 import java.util.List;
 
@@ -17,9 +21,9 @@ import java.util.List;
 public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
-    private final ImageRepository imageRepository;
     private final PasswordEncoder passwordEncoder;
     private final RoleService roleService;
+    private final BindingResultLogs bindingResultLogs;
 
     @Override
     public User getUserById(Long id) {
@@ -53,6 +57,16 @@ public class UserServiceImpl implements UserService {
         user.setAvatar(new Image(null, "images/user.jpg"));
         user.setRoles(roleService.defaultRolesSet());
         return userRepository.save(user);
+    }
+
+    @Override
+    public boolean checkUserDataBeforeReg(User user, BindingResult bindingResult, Logger logger) {
+        if (getUserByEmail(user.getEmail()) == null
+                && bindingResultLogs.checkUserFields(bindingResult, logger)) {
+            regUser(user);
+            return true;
+        }
+        return false;
     }
