+++ src/main/resources/static/js/create_new_user_page.js
@@ -70,4 +70,6 @@ async function regNewUser() {
         }
         document.getElementById("createUserResult").innerText = "Creation failed";
     }
+
+
 }
